refactor(core): moved all of the 'core' code to 'design' * split up scriptLoading.js into multiple specialized modules * created a 'code-loading' folder & put relevant code there * same for 'parameters' * modified all relevant imports
