+++ src/design/code-analysis/foo.js
