+++ src/design/rebuildSolid.js
