+++ src/design/scripLoading.js
@@ -1,53 +1,5 @@
-const fs = require('fs')
 const path = require('path')
-const getParameterDefinitionsCLI = require('./getParameterDefinitionsCLI')
-const stripBom = require('strip-bom')
-
-require.extensions['.jscad'] = function (module, filename) {
-  const content = fs.readFileSync(filename, 'utf8')
-  module._compile(stripBom(content), filename)
-}
-
-/** load an npm module from a string
- * @param  {String} src the source code of the module
- * @param  {String} filename the filename of the module
- */
-function requireFromString (src, filename) {
-  // completely ripped out of https://github.com/floatdrop/require-from-string
-  // shamefully
-  const Module = require('module')
-  const paths = Module._nodeModulePaths(path.dirname(filename))
-  const otherPaths = [path.resolve(__dirname, '../../node_modules/')]
-  const parent = module.parent
-  const m = new Module(filename, parent)
-  m.filename = filename
-  m.paths = [].concat(otherPaths).concat(paths)
-  m._compile(src, filename)
-
-  const exports = m.exports
-  parent.children && parent.children.splice(parent.children.indexOf(m), 1)
-
-  return exports
-  /* var Module = module.constructor
-  var m = new Module()
-  m._compile(src, filename)
-  return m.exports */
-}
-
-// from https://stackoverflow.com/questions/9210542/node-js-require-cache-possible-to-invalidate/16060619#16060619
-function requireUncached (moduleName) {
-  /* console.log(`removing ${moduleName} from cache`)
-  delete require.cache[require.resolve(moduleName)]
-
-  Object.keys(module.constructor._pathCache).forEach(function (cacheKey) {
-    if (cacheKey.indexOf(moduleName) > 0) {
-      delete module.constructor._pathCache[cacheKey]
-    }
-  })
-  return require(moduleName) */
-  var decache = require('decache')
-  decache(moduleName)
-}
+const getParameterDefinitionsCLI = require('./parameters/getParameterDefinitionsCLI')
 
 const doesModuleExportParameterDefiniitions = moduleToCheck => {
   return moduleToCheck && 'getParameterDefinitions' in moduleToCheck
@@ -102,4 +54,4 @@ function loadScript (scriptAsText, filePath, csgBasePath = '@jscad/csg/api') {
   return {params, paramDefinitions, scriptRootModule: scriptRootModule}
 }
 
-module.exports = {loadScript, requireUncached, requireFromString}
+module.exports = {loadScript}
