+++ src/sideEffects/fileWatcher.js
@@ -1,48 +1,8 @@
 const fs = require('fs')
-const {requireUncached} = require('../core/scripLoading')
+const requireUncached = require('../design/code-loading/requireUncached')
+const resolveDependencies = require('../design/code-loading/resolveDependencies')
 const callBackToStream = require('../utils/observable-utils/callbackToObservable')
 
- // deal with all the dependencies
-const path = require('path')
-const detective = require('detective-cjs')
-
-function resolveDependencies (options, rootPath, depth = 0) {
-  depth += 1
-  const defaults = { maxDepth: 3, relativeOnly: true }
-  options = Object.assign({}, defaults, options)
-
-  if (depth > options.maxDepth) {
-    return [rootPath]
-  }
-
-  const mainScriptAsText = fs.readFileSync(rootPath, 'utf8')
-  const rawDependencies = detective(mainScriptAsText)
-  const dependencies = Array.from(new Set(rawDependencies))
-  // console.log(`rawDependencies of ${rootPath}`, dependencies)
-  // FIXME : shamefully stolen from https://github.com/jkroso/node-resolve-module
-  function resolve (parent, name) {
-    var Module = require('module')
-    return Module._resolveFilename(name, {
-      paths: Module._nodeModulePaths(path.dirname(parent)),
-      filename: parent,
-      id: parent
-    })
-  }
-  // const dependencyAbsPaths = dependencies
-  // depPath.includes('.') : path.join(mainPath, depPath), depPath)
-  const scriptModulePaths = dependencies // [rootPath, ...dependencyAbsPaths]
-    .filter(depPath => {
-      const results = !options.relativeOnly || (options.relativeOnly && depPath.includes('.'))
-      return results
-    })
-
-  const scriptModuleAbsPaths = scriptModulePaths.map(depPath => resolve(rootPath, depPath))
-  // console.log(`dependencies of ${rootPath}`, scriptModulePaths, scriptModuleAbsPaths, depth)
-  const subResults = scriptModuleAbsPaths.map(function (scriptPath) {
-    return resolveDependencies(undefined, scriptPath, depth)
-  })
-  return [rootPath].concat(subResults)
-}
 const flatten = list => list.reduce(
   (a, b) => a.concat(Array.isArray(b) ? flatten(b) : b), []
 )
